ReportMind/Python-api/requirements.txt
pandas
openpyxl
fastapi
uvicorn
python-dotenv
langchain
langchain-chroma
langchain-core
langchain-openai